Move SfProductCard from molecules to organisms
